+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod cpu;
 pub mod devices;
 pub mod dtb;
+pub mod gdb;
 pub mod loader;
 pub mod memory;
 pub mod vm;